[procedure] use MigrateIcebergTableProcedure for migration
# Conflicts:
#	paimon-flink/paimon-flink-common/src/main/resources/META-INF/services/org.apache.paimon.factories.Factory

# Conflicts:
#	paimon-flink/paimon-flink-common/src/main/resources/META-INF/services/org.apache.paimon.factories.Factory
