+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/utils/TableMigrationUtils.java
@@ -26,7 +26,6 @@ import org.apache.paimon.iceberg.migrate.IcebergMigrator;
 import org.apache.paimon.migrate.Migrator;
 import org.apache.paimon.options.Options;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
@@ -42,28 +41,6 @@ public class TableMigrationUtils {
             String targetTableName,
             Integer parallelism,
             Map<String, String> options) {
-        return getImporter(
-                connector,
-                catalog,
-                sourceDatabase,
-                sourceTableName,
-                targetDatabase,
-                targetTableName,
-                parallelism,
-                options,
-                Collections.emptyMap());
-    }
-
-    public static Migrator getImporter(
-            String connector,
-            Catalog catalog,
-            String sourceDatabase,
-            String sourceTableName,
-            String targetDatabase,
-            String targetTableName,
-            Integer parallelism,
-            Map<String, String> options,
-            Map<String, String> icebergOptions) {
         switch (connector) {
             case "hive":
                 if (catalog instanceof CachingCatalog) {
@@ -80,19 +57,30 @@ public class TableMigrationUtils {
                         targetTableName,
                         parallelism,
                         options);
-            case "iceberg":
-                Options icebergConf = new Options(icebergOptions);
-                return new IcebergMigrator(
-                        catalog,
-                        targetDatabase,
-                        targetTableName,
-                        sourceDatabase,
-                        sourceTableName,
-                        icebergConf,
-                        parallelism);
             default:
                 throw new UnsupportedOperationException("Don't support connector " + connector);
         }
+    }
+
+    public static Migrator getIcebergImporter(
+            Catalog catalog,
+            String sourceDatabase,
+            String sourceTableName,
+            String targetDatabase,
+            String targetTableName,
+            Integer parallelism,
+            Map<String, String> options,
+            Map<String, String> icebergOptions) {
+
+        Options icebergConf = new Options(icebergOptions);
+        return new IcebergMigrator(
+                catalog,
+                targetDatabase,
+                targetTableName,
+                sourceDatabase,
+                sourceTableName,
+                icebergConf,
+                parallelism);
     }
 
     public static List<Migrator> getImporters(
